backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==44.0.0
python-multipart==0.0.18
